# Improve TakeLast performance