# Summary
In this Pull-Request, I've improved the performance of the `TakeLast/TakeRange(^*..*)` method.
#118312

# Benchmark
BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.4652)
13th Gen Intel Core i7-13700F, 1 CPU, 24 logical and 16 physical cores
.NET SDK 10.0.100-preview.6.25358.103
  [Host]     : .NET 10.0.0 (10.0.25.35903), X64 RyuJIT AVX2
  DefaultJob : .NET 10.0.0 (10.0.25.35903), X64 RyuJIT AVX2


| Method                 | Mean        | Error     | StdDev    | Gen0   | Gen1   | Allocated |
|----------------------- |------------:|----------:|----------:|-------:|-------:|----------:|
| TakeFromObjects1000    | 4,579.24 ns | 14.173 ns | 11.065 ns | 0.5341 | 0.0076 |    8480 B |
| PR_TakeFromObjects1000 | 4,057.06 ns |  9.690 ns |  8.092 ns | 0.2594 |      - |    4104 B |
| TakeFromInts1000       | 3,869.37 ns |  9.264 ns |  7.736 ns | 0.2747 |      - |    4400 B |
| PR_TakeFromInts1000    | 1,894.29 ns |  2.945 ns |  2.755 ns | 0.1335 |      - |    2096 B |
| TakeFromObjects100     |   551.21 ns |  2.114 ns |  1.874 ns | 0.0782 |      - |    1240 B |
| PR_TakeFromObjects100  |   529.36 ns |  0.869 ns |  0.771 ns | 0.0315 |      - |     504 B |
| TakeFromInts100        |   435.03 ns |  2.467 ns |  2.308 ns | 0.0472 |      - |     744 B |
| PR_TakeFromInts100     |   228.42 ns |  0.707 ns |  0.661 ns | 0.0188 |      - |     296 B |
| TakeFromObjects10      |    87.87 ns |  0.280 ns |  0.248 ns | 0.0173 |      - |     272 B |
| PR_TakeFromObjects10   |   154.01 ns |  0.256 ns |  0.239 ns | 0.0086 |      - |     136 B |
| TakeFromInts10         |    60.60 ns |  0.141 ns |  0.118 ns | 0.0142 |      - |     224 B |
| PR_TakeFromInts10      |    36.34 ns |  0.132 ns |  0.124 ns | 0.0071 |      - |     112 B |

```csharp
using BenchmarkDotNet.Attributes;


[MemoryDiagnoser]
public class TakeRangeStartIndexFromEndBenchmark
{
    IEnumerable<object?> _objects1000 = null!;

    IEnumerable<object?> _objects100 = null!;

    IEnumerable<object?> _objects10 = null!;

    IEnumerable<int> _ints1000 = null!;
    IEnumerable<int> _ints100 = null!;
    IEnumerable<int> _ints10 = null!;

    [GlobalSetup]
    public void Setup()
    {
        _objects1000 = GetEnumerable<object>(1000);
        _objects100 = GetEnumerable<object>(100);
        _objects10 = GetEnumerable<object>(10);
        _ints1000 = GetEnumerable<int>(1000);
        _ints100 = GetEnumerable<int>(100);
        _ints10 = GetEnumerable<int>(10);


        static IEnumerable<T?> GetEnumerable<T>(int count)
        {
            for (int i = 0; i < count; i++)
            {
                yield return (default(T));
            }
        }
    }


    [Benchmark]
    public void TakeFromObjects1000()
    {
        foreach (var _ in _objects1000.Take(^500..^100))
        {
        }
    }

    [Benchmark]
    public void PR_TakeFromObjects1000()
    {
        foreach (var _ in _objects1000.TakePR(^500..^100))
        {
        }
    }

    [Benchmark]
    public void TakeFromInts1000()
    {
        foreach (var _ in _ints1000.Take(^500..^100))
        {
        }
    }

    [Benchmark]
    public void PR_TakeFromInts1000()
    {
        foreach (var _ in _ints1000.TakePR(^500..^100))
        {
        }
    }


    [Benchmark]
    public void TakeFromObjects100()
    {
        foreach (var _ in _objects100.Take(^50..^10))
        {
        }
    }

    [Benchmark]
    public void PR_TakeFromObjects100()
    {
        foreach (var _ in _objects100.TakePR(^50..^10))
        {
        }
    }

    [Benchmark]
    public void TakeFromInts100()
    {
        foreach (var _ in _ints100.Take(^50..^10))
        {
        }
    }

    [Benchmark]
    public void PR_TakeFromInts100()
    {
        foreach (var _ in _ints100.TakePR(^50..^10))
        {
        }
    }

    [Benchmark]
    public void TakeFromObjects10()
    {
        foreach (var _ in _objects10.Take(^5..^1))
        {
        }
    }

    [Benchmark]
    public void PR_TakeFromObjects10()
    {
        foreach (var _ in _objects10.TakePR(^5..^1))
        {
        }
    }

    [Benchmark]
    public void TakeFromInts10()
    {
        foreach (var _ in _ints10.Take(^5..^1))
        {
        }
    }

    [Benchmark]
    public void PR_TakeFromInts10()
    {
        foreach (var _ in _ints10.TakePR(^5..^1))
        {
        }
    }
}
```