@Akeit0 2025-08-07
> As far as I understand, even though you're using a stack allocated array you still resort to allocating a final array assigned to `Source`.
> 
> I'm wondering if it would be better to always go to the ArrayPool array by default. Then, you can return the `FromEndRangeBuffer` directly from the nested method, and simply use a new `int _position` variable as an index to `yield return`.
> 
> Then just dispose the ArrayPool Array after yielding all elements.

Thanks for the review!
The reason I return the array just before yielding is that not all yielding are done synchronously, and it is possible to hold an array rent from ArrayPool. In this case, it is customary not to use rent arrays.

Example 
https://github.com/dotnet/dotnet/blob/75972a5ba730bdaf7cf3a34f528ab0f5c7f05183/src/runtime/src/libraries/System.Linq.AsyncEnumerable/src/System/Linq/ToArrayAsync.cs

By the way, in ZLinq, Enumerable is a ref struct and is almost guaranteed to be synchronous, so the last array allocation is prevented in the way you describe.

https://neuecc.medium.com/zlinq-a-zero-allocation-linq-library-for-net-1bb0a3e5c749