@skyoxZ 2025-08-08
The benchmark shows little improvement on Objects but great on Ints. Why?